add: The ability to receive a schedule for any day of the week

+++ Parser/Parser.cs
@@ -41,19 +41,21 @@ namespace ScheduleBot {
 
             if(disciplines != null) {
                 var dates = GetDates();
-                var _list = dbContext.GetDisciplinesBetweenDates(dates).ToList();
+                if(dates != null) {
+                    var _list = dbContext.GetDisciplinesBetweenDates(dates.Value).ToList();
 
-                var except = disciplines.Except(_list);
-                if(except.Any()) {
-                    AddToSchedule(except);
-                    dbContext.SaveChanges();
-                    _list = dbContext.GetDisciplinesBetweenDates(dates).ToList();
-                }
+                    var except = disciplines.Except(_list);
+                    if(except.Any()) {
+                        AddToSchedule(except);
+                        dbContext.SaveChanges();
+                        _list = dbContext.GetDisciplinesBetweenDates(dates.Value).ToList();
+                    }
 
-                except = _list.Except(disciplines);
-                if(except.Any()) {
-                    dbContext.Disciplines.RemoveRange(except);
-                    dbContext.SaveChanges();
+                    except = _list.Except(disciplines);
+                    if(except.Any()) {
+                        dbContext.Disciplines.RemoveRange(except);
+                        dbContext.SaveChanges();
+                    }
                 }
             }
 
@@ -64,75 +66,83 @@ namespace ScheduleBot {
             var completedDiscipline = dbContext.CompletedDisciplines.ToList();
 
             foreach(var discipline in list)
-                discipline.IsCompleted =(discipline.Class == DB.Entity.Type.lab && discipline.Subgroup != subgroup) | completedDiscipline.Contains(new() { Name = discipline.Name, Class = discipline.Class });
-            
+                discipline.IsCompleted = (discipline.Class == DB.Entity.Type.lab && discipline.Subgroup != subgroup) | completedDiscipline.Contains(new() { Name = discipline.Name, Class = discipline.Class });
+
             dbContext.Disciplines.AddRange(list);
         }
 
         public List<Discipline>? GetDisciplines() {
-            using(var client = new HttpClient(clientHandler, false)) {
-                #region RequestHeaders
-                client.DefaultRequestHeaders.Add("Accept", "application/json, text/javascript, */*; q=0.01");
-                client.DefaultRequestHeaders.Add("Accept-Encoding", "gzip, deflate, br");
-                client.DefaultRequestHeaders.Add("Accept-Language", "ru,en;q=0.9,en-GB;q=0.8,en-US;q=0.7");
-                client.DefaultRequestHeaders.Add("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/112.0.0.0 Safari/537.36 Edg/112.0.1722.34");
-                client.DefaultRequestHeaders.Add("X-Requested-With", "XMLHttpRequest");
-                client.DefaultRequestHeaders.Add("Referer", "https://tulsu.ru/schedule/?search=220611");
-                client.DefaultRequestHeaders.Add("Origin", "https://tulsu.ru");
-                client.DefaultRequestHeaders.Add("sec-ch-ua", "\"Chromium\";v=\"112\", \"Microsoft Edge\";v=\"112\", \"Not:A-Brand\";v=\"99\"");
-                client.DefaultRequestHeaders.Add("sec-ch-ua-mobile", "?0");
-                client.DefaultRequestHeaders.Add("sec-ch-ua-platform", "\"Windows\"");
-                client.DefaultRequestHeaders.Add("Sec-Fetch-Dest", "empty");
-                client.DefaultRequestHeaders.Add("Sec-Fetch-Mode", "cors");
-                client.DefaultRequestHeaders.Add("Sec-Fetch-Site", "same-origin");
-                client.DefaultRequestHeaders.Add("Connection", "keep-alive");
-                client.DefaultRequestHeaders.Add("Host", "tulsu.ru");
+            try {
+                using(var client = new HttpClient(clientHandler, false)) {
+                    #region RequestHeaders
+                    client.DefaultRequestHeaders.Add("Accept", "application/json, text/javascript, */*; q=0.01");
+                    client.DefaultRequestHeaders.Add("Accept-Encoding", "gzip, deflate, br");
+                    client.DefaultRequestHeaders.Add("Accept-Language", "ru,en;q=0.9,en-GB;q=0.8,en-US;q=0.7");
+                    client.DefaultRequestHeaders.Add("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/112.0.0.0 Safari/537.36 Edg/112.0.1722.34");
+                    client.DefaultRequestHeaders.Add("X-Requested-With", "XMLHttpRequest");
+                    client.DefaultRequestHeaders.Add("Referer", "https://tulsu.ru/schedule/?search=220611");
+                    client.DefaultRequestHeaders.Add("Origin", "https://tulsu.ru");
+                    client.DefaultRequestHeaders.Add("sec-ch-ua", "\"Chromium\";v=\"112\", \"Microsoft Edge\";v=\"112\", \"Not:A-Brand\";v=\"99\"");
+                    client.DefaultRequestHeaders.Add("sec-ch-ua-mobile", "?0");
+                    client.DefaultRequestHeaders.Add("sec-ch-ua-platform", "\"Windows\"");
+                    client.DefaultRequestHeaders.Add("Sec-Fetch-Dest", "empty");
+                    client.DefaultRequestHeaders.Add("Sec-Fetch-Mode", "cors");
+                    client.DefaultRequestHeaders.Add("Sec-Fetch-Site", "same-origin");
+                    client.DefaultRequestHeaders.Add("Connection", "keep-alive");
+                    client.DefaultRequestHeaders.Add("Host", "tulsu.ru");
 
-                #endregion
+                    #endregion
 
-                using(var content = new StringContent($"search_field=GROUP_P&search_value={group}", Encoding.UTF8, "application/x-www-form-urlencoded"))
-                using(HttpResponseMessage response = client.PostAsync("https://tulsu.ru/schedule/queries/GetSchedule.php", content).Result)
-                    if(response.IsSuccessStatusCode) {
-                        JArray jObject = JArray.Parse(response.Content.ReadAsStringAsync().Result);
+                    using(var content = new StringContent($"search_field=GROUP_P&search_value={group}", Encoding.UTF8, "application/x-www-form-urlencoded"))
+                    using(HttpResponseMessage response = client.PostAsync("https://tulsu.ru/schedule/queries/GetSchedule.php", content).Result)
+                        if(response.IsSuccessStatusCode) {
+                            JArray jObject = JArray.Parse(response.Content.ReadAsStringAsync().Result);
 
-                        return jObject.Select(j => new Discipline(j)).ToList();
-                    }
+                            return jObject.Select(j => new Discipline(j)).ToList();
+                        }
+                }
+            } catch(Exception) {
+                return null;
             }
+
             return null;
         }
 
-        public (DateOnly min, DateOnly max) GetDates() {
-            using(var client = new HttpClient(clientHandler, false)) {
-                #region RequestHeaders
-                client.DefaultRequestHeaders.Add("Accept", "application/json, text/javascript, */*; q=0.01");
-                client.DefaultRequestHeaders.Add("Accept-Encoding", "gzip, deflate, br");
-                client.DefaultRequestHeaders.Add("Accept-Language", "ru,en;q=0.9,en-GB;q=0.8,en-US;q=0.7");
-                client.DefaultRequestHeaders.Add("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/112.0.0.0 Safari/537.36 Edg/112.0.1722.34");
-                client.DefaultRequestHeaders.Add("X-Requested-With", "XMLHttpRequest");
-                client.DefaultRequestHeaders.Add("Referer", "https://tulsu.ru/schedule/?search=220611");
-                client.DefaultRequestHeaders.Add("Origin", "https://tulsu.ru");
-                client.DefaultRequestHeaders.Add("sec-ch-ua", "\"Chromium\";v=\"112\", \"Microsoft Edge\";v=\"112\", \"Not:A-Brand\";v=\"99\"");
-                client.DefaultRequestHeaders.Add("sec-ch-ua-mobile", "?0");
-                client.DefaultRequestHeaders.Add("sec-ch-ua-platform", "\"Windows\"");
-                client.DefaultRequestHeaders.Add("Sec-Fetch-Dest", "empty");
-                client.DefaultRequestHeaders.Add("Sec-Fetch-Mode", "cors");
-                client.DefaultRequestHeaders.Add("Sec-Fetch-Site", "same-origin");
-                client.DefaultRequestHeaders.Add("Connection", "keep-alive");
-                client.DefaultRequestHeaders.Add("Host", "tulsu.ru");
+        public (DateOnly min, DateOnly max)? GetDates() {
+            try {
+                using(var client = new HttpClient(clientHandler, false)) {
+                    #region RequestHeaders
+                    client.DefaultRequestHeaders.Add("Accept", "application/json, text/javascript, */*; q=0.01");
+                    client.DefaultRequestHeaders.Add("Accept-Encoding", "gzip, deflate, br");
+                    client.DefaultRequestHeaders.Add("Accept-Language", "ru,en;q=0.9,en-GB;q=0.8,en-US;q=0.7");
+                    client.DefaultRequestHeaders.Add("User-Agent", "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/112.0.0.0 Safari/537.36 Edg/112.0.1722.34");
+                    client.DefaultRequestHeaders.Add("X-Requested-With", "XMLHttpRequest");
+                    client.DefaultRequestHeaders.Add("Referer", "https://tulsu.ru/schedule/?search=220611");
+                    client.DefaultRequestHeaders.Add("Origin", "https://tulsu.ru");
+                    client.DefaultRequestHeaders.Add("sec-ch-ua", "\"Chromium\";v=\"112\", \"Microsoft Edge\";v=\"112\", \"Not:A-Brand\";v=\"99\"");
+                    client.DefaultRequestHeaders.Add("sec-ch-ua-mobile", "?0");
+                    client.DefaultRequestHeaders.Add("sec-ch-ua-platform", "\"Windows\"");
+                    client.DefaultRequestHeaders.Add("Sec-Fetch-Dest", "empty");
+                    client.DefaultRequestHeaders.Add("Sec-Fetch-Mode", "cors");
+                    client.DefaultRequestHeaders.Add("Sec-Fetch-Site", "same-origin");
+                    client.DefaultRequestHeaders.Add("Connection", "keep-alive");
+                    client.DefaultRequestHeaders.Add("Host", "tulsu.ru");
 
-                #endregion
+                    #endregion
 
-                using(var content = new StringContent($"search_value={group}", Encoding.UTF8, "application/x-www-form-urlencoded"))
-                using(HttpResponseMessage response = client.PostAsync("https://tulsu.ru/schedule/queries/GetDates.php", content).Result)
-                    if(response.IsSuccessStatusCode) {
-                        JObject jObject = JObject.Parse(response.Content.ReadAsStringAsync().Result);
+                    using(var content = new StringContent($"search_value={group}", Encoding.UTF8, "application/x-www-form-urlencoded"))
+                    using(HttpResponseMessage response = client.PostAsync("https://tulsu.ru/schedule/queries/GetDates.php", content).Result)
+                        if(response.IsSuccessStatusCode) {
+                            JObject jObject = JObject.Parse(response.Content.ReadAsStringAsync().Result);
 
-                        return (DateOnly.Parse(jObject.Value<string>("MIN_DATE") ?? throw new NullReferenceException("MIN_DATE")),
-                                DateOnly.Parse(jObject.Value<string>("MAX_DATE") ?? throw new NullReferenceException("MAX_DATE")));
-                    }
+                            return (DateOnly.Parse(jObject.Value<string>("MIN_DATE") ?? throw new NullReferenceException("MIN_DATE")),
+                                    DateOnly.Parse(jObject.Value<string>("MAX_DATE") ?? throw new NullReferenceException("MAX_DATE")));
+                        }
+                }
+            } catch(Exception) {
+                return null;
             }
-
-            throw new NullReferenceException("Dates");
+            return null;
         }
     }
 }
